backend/requirements.txt
fastapi
uvicorn
numpy
scikit-learn
spacy
PyPDF2
python-docx
requests 